Undo float32

+++ pf_localization/pf_localization/pf_node.py
@@ -184,7 +184,7 @@ class ParticleFilterNode(Node):
         #Use this to get a measurement and an uncertainty
 
         #get a measurement based on the current pose of the lidar
-        measurement = np.zeros((self.lidar_resolution, 1), dtype=np.float32)
+        measurement = np.zeros((self.lidar_resolution, 1))
         
         #determine where the pose is on the map
         map_pos = (np.array(pose[:2]) - self.origin[:2]) / self.map_resolution
@@ -195,7 +195,7 @@ class ParticleFilterNode(Node):
             ray_distance = 0
             
             #unit vector in the direction of the ray
-            angle_unit_vector = np.array([cos(measurement_angle), sin(measurement_angle)], dtype=np.float32)
+            angle_unit_vector = np.array([cos(measurement_angle), sin(measurement_angle)])
 
             while(not collision_found):
                 #increment the ray distance
@@ -224,11 +224,11 @@ class ParticleFilterNode(Node):
         current_ray_pos = map_pos + unit_vector * current_dist
 
         #determine the current cell
-        current_cell = np.array([floor(current_ray_pos[0]), floor(current_ray_pos[1])], dtype=np.float32)
+        current_cell = np.array([floor(current_ray_pos[0]), floor(current_ray_pos[1])])
         
         #determine the next cell in the direction
-        next_cell_x = current_cell + np.array([unit_vector[0] / abs(unit_vector[0]), 0], dtype=np.float32)
-        next_cell_y = current_cell + np.array([0, unit_vector[1] / abs(unit_vector[1])], dtype=np.float32) 
+        next_cell_x = current_cell + np.array([unit_vector[0] / abs(unit_vector[0]), 0])
+        next_cell_y = current_cell + np.array([0, unit_vector[1] / abs(unit_vector[1])]) 
 
         #determine if it shorter to the next cell to increment to the x bound or y bound
         dist_x = (next_cell_x[0]  - current_ray_pos[0]) / unit_vector[0]
@@ -258,7 +258,7 @@ class ParticleFilterNode(Node):
 
         #use this to check if the estimate is in collision with anything
 
-        map_pos = (np.array(pose[:2], dtype=np.float32) - self.origin[:2]) / self.map_resolution
+        map_pos = (np.array(pose[:2]) - self.origin[:2]) / self.map_resolution
 
         return self.check_collision_map_frame(map_pos)
 
@@ -266,7 +266,7 @@ class ParticleFilterNode(Node):
     def get_img_index_from_pos(self, pos):
 
         #transform to map frame
-        map_pos = (np.array(pos, dtype=np.float32) - self.origin[:2]) / self.map_resolution
+        map_pos = (np.array(pos) - self.origin[:2]) / self.map_resolution
 
         if(map_pos[0] < 0):
             map_pos[0] = 0
